requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database & Caching
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# AI & ML (OpenAI Only)
openai==1.3.7
langchain==0.0.350
langgraph==0.0.20

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78

# Audio Processing
pydub==0.25.1
speechrecognition==3.10.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
